refactor: 예약 시간 관련 기능 통합
