+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -3,12 +3,15 @@ package roomescape.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
+import roomescape.domain.dto.BookResponse;
 import roomescape.domain.dto.ReservationTimeRequest;
 import roomescape.domain.dto.ReservationTimeResponse;
 import roomescape.exception.DeleteNotAllowException;
@@ -68,6 +71,15 @@ class ReservationTimeServiceTest {
         //when, then
         assertThatThrownBy(() -> service.delete(1L)).isInstanceOf(DeleteNotAllowException.class);
         assertThat(getTimeSlotSize()).isEqualTo(initialSize);
+    }
 
+    @DisplayName("예약 가능한 시간 목록들을 반환한다.")
+    @Test
+    void given_when_findAvailableBookList_thenReturnBookResponse() {
+        //when
+        final List<BookResponse> bookResponses = service.findAvailableBookList(LocalDate.parse("2099-05-08"), 1L)
+                .getData();
+        //then
+        assertThat(bookResponses.size()).isEqualTo(4);
     }
 }
